+++ src/components/Header.tsx
@@ -11,14 +11,11 @@ export default function Header({ left, right }: { left: ReactNode; right?: React
 }
 
 const headerCss = css({
-  position: 'sticky',
-  top: '0%',
   height: '64px',
   width: '430px',
   backgroundColor: 'black',
   padding: '24px',
   display: 'flex',
   justifyContent: 'space-between',
-  alignItems: 'center',
-  zIndex: 100
+  alignItems: 'center'
 });
